refactor: Simplify layout and improve responsiveness in Cource and Comparison components by removing unnecessary breakpoints and enhancing text formatting
