+++ src/components/CareerPath/CareerPath.tsx
@@ -1,13 +1,7 @@
-'use client';
-
-import useIsPc from '@/hooks/useIsPc';
-
 import styles from './CareerPath.module.css';
 import CareerPathSlider from './CareerPathSlider/CareerPathSlider';
 
 export default function CareerPath() {
-  const isPc = useIsPc();
-
   return (
     <div className={styles.CareerPath}>
       <div className={styles.CareerPath__Header}>
@@ -15,7 +9,11 @@ export default function CareerPath() {
         <h2 className={styles.CareerPath__Heading}>卒業生の進路</h2>
         <div className={styles.CareerPath__DetailButtonContainer}>
           <a href="#" className={styles.CareerPath__DetailButton}>
-            <span className={styles.CareerPath__DetailButtonText}>インタビューの{!isPc && <br />}詳細を見る</span>
+            <span className={styles.CareerPath__DetailButtonText}>
+              インタビューの
+              <br className="br-sp" />
+              詳細を見る
+            </span>
             <svg width="11" height="11" fill="none">
               <use href="#rightArrowWhite"></use>
             </svg>
