+++ src/components/Comparison/Comparison.tsx
@@ -1,14 +1,9 @@
-'use client';
-
-import useIsPc from '@/hooks/useIsPc';
-
 import Logo from '../Logo/Logo';
 
 import styles from './Comparison.module.css';
 import ComparisonTable from './ComparisonTable/ComparisonTable';
 
 export default function Comparison() {
-  const isPc = useIsPc();
   return (
     <div className={styles.Comparison}>
       <div className={styles.Comparison__Container}>
@@ -16,7 +11,7 @@ export default function Comparison() {
         <h2 className={styles.Comparison__Heading}>
           <Logo fill="#000" />
           と他スクールとの
-          {!isPc && <br />}
+          <br className="br-sp" />
           比較
         </h2>
