+++ src/components/Cource/Cource.module.css
@@ -46,82 +46,8 @@
   display: inline-flex;
 }
 
-/* カードスライダー */
 .Cource__Slider {
-  position: relative;
-  perspective: 1000px;
   margin-bottom: 46px;
-  &::before {
-    content: '';
-    display: block;
-    width: 100vw;
-    height: 100%;
-    background-color: #fff;
-    position: absolute;
-    top: 50%;
-    left: -32px;
-    z-index: 0;
-    border-radius: 20px 20px 0 0;
-  }
-}
-
-.Cource__CardStack {
-  display: grid;
-  grid-template-columns: 1fr;
-  grid-template-rows: 1fr;
-  max-width: 307px;
-  margin: 0 auto;
-  transform-style: preserve-3d;
-}
-
-.Cource__Card {
-  grid-column: 1;
-  grid-row: 1;
-  transition: all 0.6s cubic-bezier(0.4, 0, 0.2, 1);
-  transform-style: preserve-3d;
-}
-
-/* 位置（順番）によって角度とY軸が変わる */
-.Cource__Card--position0 {
-  transform: translateY(0px) translateZ(0px) rotate(0deg);
-  z-index: 3;
-}
-
-.Cource__Card--position1 {
-  transform: translateY(-30px) translateZ(-10px) rotate(3.6deg);
-  z-index: 2;
-}
-
-.Cource__Card--position2 {
-  transform: translateY(10px) translateZ(-20px) rotate(-2.8deg);
-  z-index: 1;
-}
-
-/* ナビゲーションボタン */
-.Cource__NavButton {
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  background-color: var(--text);
-  color: #fff;
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 20px;
-  z-index: 10;
-}
-
-.Cource__NavButton:first-of-type {
-  left: -20px;
-}
-
-.Cource__NavButton:last-of-type {
-  right: -20px;
 }
 
 .Cource__ReskillBanner {
@@ -243,52 +169,7 @@
   }
 
   .Cource__Slider {
-    perspective: none; /* 3D効果を無効化 */
-    padding: 0;
-    &::before {
-      border-radius: 40px 40px 0 0;
-      width: calc(100% + 180px);
-      top: 55%;
-      left: 50%;
-      transform: translateX(-50%);
-    }
-  }
-
-  .Cource__CardStack {
-    display: flex;
-    justify-content: space-between;
-    max-width: 1030px;
-    transform-style: inherit;
-  }
-
-  .Cource__Card {
-    position: static;
-    opacity: 1; /* すべて表示 */
-    z-index: auto;
-    transform-style: inherit;
-  }
-
-  .Cource__Card--position0 {
-    order: 1;
-    transform: none; /* 3D変形をリセット */
-    flex-basis: 360px;
-    z-index: 1;
-  }
-
-  .Cource__Card--position1 {
-    order: 2;
-    transform: rotate(10deg) translateY(20px);
-    flex-basis: 340px;
-  }
-
-  .Cource__Card--position2 {
-    order: 0;
-    transform: rotate(-12deg) translateY(34px);
-    flex-basis: 340px;
-  }
-
-  .Cource__NavButton {
-    display: none;
+    margin-bottom: 60px;
   }
 
   .Cource__ReskillBanner {
