+++ src/components/Cource/Cource.tsx
@@ -1,66 +1,12 @@
-'use client';
-
-import { useRef, useState } from 'react';
-
-import { COURSES } from '@/constants/courseData';
-import useIsPc from '@/hooks/useIsPc';
-
 import Logo from '../Logo/Logo';
 
 import styles from './Cource.module.css';
-import CourseCard from './CourseCard/CourseCard';
+import CourseSlider from './CourseSlider/CourseSlider';
 import FutureItem from './FutureItem/FutureItem';
 import ReskillBannerSection from './ReskillBannerSection/ReskillBannerSection';
 import Timeline from './Timeline/Timeline';
 
 export default function Cource() {
-  const [activeIndex, setActiveIndex] = useState(0);
-  const [isAnimating, setIsAnimating] = useState(false);
-  const touchStartX = useRef(0);
-  const touchEndX = useRef(0);
-
-  const handleNext = () => {
-    if (isAnimating) {
-      return;
-    }
-    setIsAnimating(true);
-    setActiveIndex(prev => (prev + 1) % COURSES.length);
-    setTimeout(() => setIsAnimating(false), 600);
-  };
-
-  const handlePrev = () => {
-    if (isAnimating) {
-      return;
-    }
-    setIsAnimating(true);
-    setActiveIndex(prev => (prev - 1 + COURSES.length) % COURSES.length);
-    setTimeout(() => setIsAnimating(false), 600);
-  };
-
-  const handleTouchStart = (e: React.TouchEvent) => {
-    touchStartX.current = e.targetTouches[0].clientX;
-  };
-
-  const handleTouchMove = (e: React.TouchEvent) => {
-    touchEndX.current = e.targetTouches[0].clientX;
-  };
-
-  const handleTouchEnd = () => {
-    if (!touchStartX.current || !touchEndX.current) {
-      return;
-    }
-
-    const distance = touchStartX.current - touchEndX.current;
-    const isLeftSwipe = distance > 50;
-    const isRightSwipe = distance < -50;
-
-    if (isLeftSwipe) {
-      handleNext();
-    } else if (isRightSwipe) {
-      handlePrev();
-    }
-  };
-
   // 未来の項目データ
   const futureItems = [
     { number: '01', text: '8-12ヶ月でITエンジニアへ' },
@@ -69,8 +15,6 @@ export default function Cource() {
     { number: '04', text: 'キャリアや経済不安からの脱却' },
     { number: '05', text: 'さまざまなキャリアの選択肢', subText: '（リモートワーク、有名企業、フリーランス、海外、など）' },
   ];
-
-  const isPc = useIsPc();
 
   return (
     <div className={styles.Cource}>
@@ -81,7 +25,11 @@ export default function Cource() {
           <h2 className={styles.Cource__Heading}>コース紹介</h2>
           <div className={styles.Cource__DetailButtonContainer}>
             <a href="#" className={styles.Cource__DetailButton}>
-              <span className={styles.Cource__DetailButtonText}>コースの{!isPc && <br />}詳細を見る</span>
+              <span className={styles.Cource__DetailButtonText}>
+                コースの
+                <br className="br-sp" />
+                詳細を見る
+              </span>
               <svg width={11} height={11} fill="none">
                 <use href="#rightArrowWhite" />
               </svg>
@@ -91,31 +39,7 @@ export default function Cource() {
 
         {/* カードスライダーセクション */}
         <div className={styles.Cource__Slider}>
-          <div className={styles.Cource__CardStack} onTouchStart={handleTouchStart} onTouchMove={handleTouchMove} onTouchEnd={handleTouchEnd}>
-            {COURSES.map((course, index) => {
-              const position = (index - activeIndex + COURSES.length) % COURSES.length;
-
-              return (
-                <div key={index} className={`${styles.Cource__Card} ${styles[`Cource__Card--${course.type}`]} ${styles[`Cource__Card--position${position}`]}`}>
-                  <CourseCard {...course} />
-                </div>
-              );
-            })}
-            <button className={styles.Cource__NavButton} onClick={handlePrev} aria-label="前のコース">
-              <span>
-                <svg width="48" height="48" viewBox="0 0 48 48" fill="none">
-                  <use href="#circleLeftAllowBlue" />
-                </svg>
-              </span>
-            </button>
-            <button className={styles.Cource__NavButton} onClick={handleNext} aria-label="次のコース">
-              <span>
-                <svg width="48" height="48" viewBox="0 0 48 48" fill="none">
-                  <use href="#circleRightAllowBlue" />
-                </svg>
-              </span>
-            </button>
-          </div>
+          <CourseSlider />
         </div>
 
         {/* Reスキル講座バナー */}
